+++ selfdrive/car/gm/interface.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-from math import fabs
+from math import fabs, sin
 from cereal import car
 from common.numpy_fast import interp
 from common.realtime import sec_since_boot
@@ -18,31 +18,40 @@ FOLLOW_AGGRESSION = 0.15 # (Acceleration/Decel aggression) Lower is more aggress
 _A_CRUISE_MIN_V_SPORT = [-2.0, -2.2, -2.0, -1.5, -1.0]
 _A_CRUISE_MIN_V_FOLLOWING = [-3.0, -2.5, -2.0, -1.5, -1.0]
 _A_CRUISE_MIN_V = [-1.0, -1.2, -1.0, -0.7, -0.5]
+_A_CRUISE_MIN_V_ECO = [-0.7, -0.8, -0.7, -0.6, -0.5]
 _A_CRUISE_MIN_BP = [0., 5., 10., 20., 55.]
 
 # need fast accel at very low speed for stop and go
 # make sure these accelerations are smaller than mpc limits
+_A_CRUISE_MAX_V_CREEP = [0.22, .22, .2, .2, .2]
+_A_CRUISE_MAX_V_ECO = [0.7, 0.7, 0.7, 0.5, 0.4]
 _A_CRUISE_MAX_V = [1.2, 1.4, 1.2, 0.9, 0.7]
 _A_CRUISE_MAX_V_SPORT = [2.2, 2.4, 2.2, 1.1, 0.9]
 _A_CRUISE_MAX_V_FOLLOWING = [1.6, 1.8, 1.6, .9, .7]
 _A_CRUISE_MAX_BP = [0., 5., 10., 20., 55.]
 
-_A_CRUISE_MIN_V_MODE_LIST = [_A_CRUISE_MIN_V, _A_CRUISE_MIN_V_SPORT]
-_A_CRUISE_MAX_V_MODE_LIST = [_A_CRUISE_MAX_V, _A_CRUISE_MAX_V_SPORT]
+_A_CRUISE_MIN_V_MODE_LIST = [_A_CRUISE_MIN_V, _A_CRUISE_MIN_V_SPORT, _A_CRUISE_MIN_V_ECO, _A_CRUISE_MIN_V_ECO]
+_A_CRUISE_MAX_V_MODE_LIST = [_A_CRUISE_MAX_V, _A_CRUISE_MAX_V_SPORT, _A_CRUISE_MAX_V_ECO, _A_CRUISE_MAX_V_CREEP]
 
 # revert to stock max negative accel based on relative lead velocity
 _A_MIN_V_STOCK_FACTOR_BP = [-5. * CV.MPH_TO_MS, 1. * CV.MPH_TO_MS]
 _A_MIN_V_STOCK_FACTOR_V = [0., 1.]
 
+# increase/decrease max accel based on vehicle pitch
+INCLINE_ACCEL_OFFSET = 0.22 # [m/s^2] desired acceleration must be at least this much more than g acceleration
+INCLINE_ACCEL_SPEED_OFFSET_BP = [i * CV.MPH_TO_MS for i in [25., 45]] # [mph] lookup speeds for additional offset
+INCLINE_ACCEL_SPEED_OFFSET_V = [0., 0.2] # [m/s^2] additional offset based on speed
+DECLINE_ACCEL_FACTOR = 0.5 # this factor of g accel is used to lower max accel limit so you don't floor it downhill
+DECLINE_ACCEL_MIN = 0.2 # [m/s^2] don't decrease acceleration limit due to decline below this total value
+
 def calc_cruise_accel_limits(v_ego, following, accelMode):
-  if following:
+  if following and accelMode < 2:
     a_cruise_min = interp(v_ego, _A_CRUISE_MIN_BP, _A_CRUISE_MIN_V_FOLLOWING)
     a_cruise_max = interp(v_ego, _A_CRUISE_MAX_BP, _A_CRUISE_MAX_V_FOLLOWING)
   else:
     a_cruise_min = interp(v_ego, _A_CRUISE_MIN_BP, _A_CRUISE_MIN_V_MODE_LIST[accelMode])
     a_cruise_max = interp(v_ego, _A_CRUISE_MAX_BP, _A_CRUISE_MAX_V_MODE_LIST[accelMode])
   return [a_cruise_min, a_cruise_max]
-
 
 ButtonType = car.CarState.ButtonEvent.Type
 EventName = car.CarEvent.EventName
@@ -56,8 +65,18 @@ class CarInterface(CarInterfaceBase):
   def get_pid_accel_limits(CP, current_speed, cruise_speed, CI = None):
     following = CI.CS.coasting_lead_d > 0. and CI.CS.coasting_lead_d < 45.0 and CI.CS.coasting_lead_v > current_speed
     accel_limits = calc_cruise_accel_limits(current_speed, following, CI.CS.accel_mode)
+    
+    # decrease min accel as necessary based on lead conditions
     stock_min_factor = interp(current_speed - CI.CS.coasting_lead_v, _A_MIN_V_STOCK_FACTOR_BP, _A_MIN_V_STOCK_FACTOR_V) if CI.CS.coasting_lead_d > 0. else 0.
     accel_limits[0] = stock_min_factor * CI.params.ACCEL_MIN + (1. - stock_min_factor) * accel_limits[0]
+    
+    # decrease/increase max accel based on vehicle pitch
+    g_accel = 9.81 * sin(CI.CS.pitch)
+    if g_accel > 0.:
+      accel_limits[1] = max(accel_limits[1], g_accel + INCLINE_ACCEL_OFFSET + interp(current_speed, INCLINE_ACCEL_SPEED_OFFSET_BP, INCLINE_ACCEL_SPEED_OFFSET_V))
+    else:
+      accel_limits[1] = max(DECLINE_ACCEL_MIN, accel_limits[1] + g_accel * DECLINE_ACCEL_FACTOR)
+      
     return [max(CI.params.ACCEL_MIN, accel_limits[0]), min(accel_limits[1], CI.params.ACCEL_MAX)]
 
   # Volt determined by iteratively plotting and minimizing error for f(angle, speed) = steer.